Computes the content of a DT::datatable for a correction set

Populate a DT table according to selected items with a list of grids (or "all") or according to a selected grid with a list of items (or "all")

populate_table(
  items,
  grids = "all",
  context,
  reorder = (length(items) == 1),
  highlight = FALSE,
  max_lines = 30L,
  on_github = TRUE
)

Arguments

items

The items of the evaluation grid to display in the table, usually either one item, or "all" for everything

grids

The evaluation grids to display, usually either "all" if only one item, or one grid if "all" items

context

A context object as computed by create_context().

reorder

Should the rows in the table be reordered by similarities? This is usually TRUE for a single items, or FALSE otherwise (and it is computed as such by default).

highlight

Syntax highlighting for code (slow, thus FALSE by default)

max_lines

The maximum number of content lines that are displayed (truncate very long contents).

on_github

Should the links point to the GitHub repository or to the local files (default is TRUE)?

Value

A data frame with the content to be displayed in a DT::datatable object.